+++ Flower_shop/Migrations/20221020143106_AllMigrations.cs
@@ -50,13 +50,13 @@ namespace Flower_shop.Migrations
                 oldClrType: typeof(string),
                 oldType: "nvarchar(max)");
 
-            migrationBuilder.AlterColumn<int>(
+            migrationBuilder.AlterColumn<long>(
                 name: "TypeProductId",
                 table: "Products",
-                type: "int",
+                type: "bigint",
                 nullable: true,
-                oldClrType: typeof(int),
-                oldType: "int");
+                oldClrType: typeof(long),
+                oldType: "bigint");
 
             migrationBuilder.AlterColumn<string>(
                 name: "Price",
@@ -154,14 +154,14 @@ namespace Flower_shop.Migrations
                 oldType: "nvarchar(max)",
                 oldNullable: true);
 
-            migrationBuilder.AlterColumn<int>(
+            migrationBuilder.AlterColumn<long>(
                 name: "TypeProductId",
                 table: "Products",
-                type: "int",
+                type: "bigint",
                 nullable: false,
                 defaultValue: 0,
-                oldClrType: typeof(int),
-                oldType: "int",
+                oldClrType: typeof(long),
+                oldType: "bigint",
                 oldNullable: true);
 
             migrationBuilder.AlterColumn<string>(
