+++ Flower_shop/Migrations/WebDbContextModelSnapshot.cs
@@ -24,11 +24,11 @@ namespace Flower_shop.Migrations
 
             modelBuilder.Entity("Flower_shop.EfStuff.DbModels.Color", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
 
                     b.Property<string>("AssignmentName")
                         .HasColumnType("nvarchar(max)");
@@ -43,11 +43,11 @@ namespace Flower_shop.Migrations
 
             modelBuilder.Entity("Flower_shop.EfStuff.DbModels.Image", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
 
                     b.Property<int>("Block")
                         .HasColumnType("int");
@@ -71,11 +71,11 @@ namespace Flower_shop.Migrations
 
             modelBuilder.Entity("Flower_shop.EfStuff.DbModels.Product", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
 
                     b.Property<string>("ImageName")
                         .HasColumnType("nvarchar(max)");
@@ -89,8 +89,8 @@ namespace Flower_shop.Migrations
                     b.Property<string>("Price")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int?>("TypeProductId")
-                        .HasColumnType("int");
+                    b.Property<long?>("TypeProductId")
+                        .HasColumnType("bigint");
 
                     b.HasKey("Id");
 
@@ -101,11 +101,11 @@ namespace Flower_shop.Migrations
 
             modelBuilder.Entity("Flower_shop.EfStuff.DbModels.TypeProduct", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
 
                     b.Property<string>("Name")
                         .HasColumnType("nvarchar(max)");
@@ -117,11 +117,11 @@ namespace Flower_shop.Migrations
 
             modelBuilder.Entity("Flower_shop.EfStuff.DbModels.User", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
 
                     b.Property<string>("Email")
                         .HasColumnType("nvarchar(max)");
